# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/uploads": {
      "post": {
        "operationId": "UploadsController_uploadFile",
        "parameters": [
          {
            "name": "folder",
            "required": false,
            "in": "query",
            "description": "Целевая папка для загрузки файла. Если не указана, используется папка по умолчанию (например, 'common-uploads'). Допустимые символы: буквы, цифры, дефис, подчеркивание. Максимальная длина: 50 символов.",
            "schema": {
              "example": "avatars",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Файл для загрузки (max 5MB, типы: .png, .jpg, .jpeg, .webp)",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Файл успешно загружен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "example": "https://storage.yandexcloud.net/your-bucket/folder/image.jpg"
                    },
                    "fileKey": {
                      "type": "string",
                      "example": "folder/image.jpg"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос, файл не предоставлен, превышен размер или неверный тип файла"
          },
          "401": {
            "description": "Пользователь не авторизован"
          },
          "500": {
            "description": "Внутренняя ошибка сервера при загрузке файла"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Загрузить файл в указанную папку (или папку по умолчанию)",
        "tags": [
          "uploads"
        ]
      }
    },
...
  },
  "info": {
    "title": "Grape Shop API",
    "description": "API для интернет-магазина саженцев винограда",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}